dem2fds source: read fire data from a text file, code clean up, remove IOadf routine
