+++ Source/dem2fds/options.h
@@ -17,6 +17,4 @@
 #define PROGVERSION "1.0.1"
 #endif
 
-//#define pp_ADF // option to parse adf files
-
 #endif
